🚀 feat(orders): implement order details view and enhance orders management
- Added `OrderDetails` component to display detailed information for individual orders, improving user experience during order review.
- Integrated new hooks and actions for fetching and managing order data, including pagination and search functionality in the `OrdersList`.
- Enhanced the `useOrders` hook to support fetching all orders with filters, improving data retrieval efficiency.
- Updated Redux slice to manage orders list state, including loading and error handling, ensuring a robust data management process.
- Improved UI components for better interaction and feedback during order management tasks.
