+++ src/App.tsx
@@ -12,6 +12,7 @@ import CustomerDetails from "@/pages/CustomerDetails";
 import TransactionTypes from "@/pages/TransactionTypes";
 import LocationManagement from "@/pages/LocationManagement";
 import OrdersList from "@/pages/OrdersList";
+import OrderDetails from "@/pages/OrderDetails";
 import CreateOrder from "@/pages/CreateOrder";
 import VendorsList from "@/pages/VendorsList";
 import VendorDetails from "@/pages/VendorDetails";
@@ -29,7 +30,7 @@ function Router() {
         <Route path="/" component={Overview} />
         <Route path="/activities" component={ActivityManagement} />
         <Route path="/orders" component={OrdersList} />
-        <Route path="/orders/:id" component={OrdersList} />
+        <Route path="/orders/:id" component={OrderDetails} />
         <Route path="/create-order" component={CreateOrder} />
         <Route path="/price-lists" component={PriceLists} />
         <Route path="/price-lists/:id" component={PriceListDetails} />
